[feature/#50] change the path for an auth page
