+++ src/app/dev/auth/authStyles.module.css
@@ -1,4 +1,4 @@
-@reference '../globals.css';
+@reference '../../globals.css';
 
 .label {
 	@apply mt-3;
